Move movableUnitsFilter, colocate with MovePanel, the only class that uses it.
