+++ game-core/src/main/java/games/strategy/triplea/ui/panel/move/MovableUnitsFilter.java
@@ -1,4 +1,4 @@
-package games.strategy.triplea.util;
+package games.strategy.triplea.ui.panel.move;
 
 import com.google.common.base.Preconditions;
 import games.strategy.engine.data.GameData;
@@ -16,6 +16,7 @@ import games.strategy.triplea.delegate.UndoableMove;
 import games.strategy.triplea.delegate.UnitComparator;
 import games.strategy.triplea.delegate.data.MoveValidationResult;
 import games.strategy.triplea.delegate.data.MustMoveWithDetails;
+import games.strategy.triplea.util.TransportUtils;
 import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -31,7 +32,7 @@ import org.triplea.java.collections.CollectionUtils;
 /**
  * Utility class for filtering a unit list to the subset of those units that can move on a route.
  */
-public final class MovableUnitsFilter {
+final class MovableUnitsFilter {
   /** The result of the filter operation. */
   @Getter
   public static class FilterOperationResult {
